+++ projects/hipcub/benchmark/benchmark_device_memory.cpp
@@ -77,12 +77,12 @@ struct operation<custom_operation, T, ItemsPerThread, BlockSize>
                            T (&input)[ItemsPerThread],
                            [[maybe_unused]] T* global_mem_output = nullptr) const
     {
-        _CCCL_PRAGMA_UNROLL_FULL()
+        HIPCUB_PRAGMA_UNROLL_FULL()
         for(unsigned int i = 0; i < ItemsPerThread; i++)
         {
             input[i]                       = input[i] + 666;
             constexpr unsigned int repeats = 30;
-            _CCCL_PRAGMA_UNROLL_FULL()
+            HIPCUB_PRAGMA_UNROLL_FULL()
             for(unsigned int j = 0; j < repeats; j++)
             {
                 input[i] = input[i] * (input[j % ItemsPerThread]);
@@ -123,7 +123,7 @@ struct operation<atomics_no_collision, T, ItemsPerThread, BlockSize>
     {
         const unsigned int index
             = threadIdx.x * ItemsPerThread + blockIdx.x * blockDim.x * ItemsPerThread;
-        _CCCL_PRAGMA_UNROLL_FULL()
+        HIPCUB_PRAGMA_UNROLL_FULL()
         for(unsigned int i = 0; i < ItemsPerThread; i++)
         {
             atomicAdd(&global_mem_output[index + i], T(666));
@@ -144,7 +144,7 @@ struct operation<atomics_inter_warp_collision, T, ItemsPerThread, BlockSize>
     {
         const unsigned int index
             = (threadIdx.x % warpSize) * ItemsPerThread + blockIdx.x * blockDim.x * ItemsPerThread;
-        _CCCL_PRAGMA_UNROLL_FULL()
+        HIPCUB_PRAGMA_UNROLL_FULL()
         for(unsigned int i = 0; i < ItemsPerThread; i++)
         {
             atomicAdd(&global_mem_output[index + i], T(666));
@@ -164,7 +164,7 @@ struct operation<atomics_inter_block_collision, T, ItemsPerThread, BlockSize>
                            T* global_mem_output = nullptr)
     {
         const unsigned int index = threadIdx.x * ItemsPerThread;
-        _CCCL_PRAGMA_UNROLL_FULL()
+        HIPCUB_PRAGMA_UNROLL_FULL()
         for(unsigned int i = 0; i < ItemsPerThread; i++)
         {
             atomicAdd(&global_mem_output[index + i], T(666));