+++ projects/hipcub/hipcub/include/hipcub/backend/rocprim/block/block_merge_sort.hpp
@@ -412,7 +412,7 @@ public:
     // each thread has sorted keys
     // merge sort keys in shared memory
     //
-    _CCCL_PRAGMA_UNROLL_FULL()
+    HIPCUB_PRAGMA_NOUNROLL()
     for (int target_merged_threads_number = 2;
          target_merged_threads_number <= NUM_THREADS;
          target_merged_threads_number *= 2)
@@ -424,7 +424,7 @@ public:
 
       // store keys in shmem
       //
-      _CCCL_PRAGMA_UNROLL_FULL()
+      HIPCUB_PRAGMA_NOUNROLL()
       for (int item = 0; item < ITEMS_PER_THREAD; ++item)
       {
         int idx                       = ITEMS_PER_THREAD * linear_tid + item;
@@ -483,7 +483,7 @@ public:
 
         // store keys in shmem
         //
-        _CCCL_PRAGMA_UNROLL_FULL()
+        HIPCUB_PRAGMA_NOUNROLL()
         for (int item = 0; item < ITEMS_PER_THREAD; ++item)
         {
           int idx = ITEMS_PER_THREAD * linear_tid + item;
@@ -494,7 +494,7 @@ public:
 
         // gather items from shmem
         //
-        _CCCL_PRAGMA_UNROLL_FULL()
+        HIPCUB_PRAGMA_NOUNROLL()
         for (int item = 0; item < ITEMS_PER_THREAD; ++item)
         {
           items[item] = temp_storage.items_shared[indices[item]];