+++ projects/hipcub/hipcub/include/hipcub/config.hpp
@@ -256,12 +256,12 @@ END_HIPCUB_NAMESPACE
 #endif // HIPCUB_TRAIT
 
 // This API needs to be deprecated once libhipcxx is available.
-#if !defined(_CCCL_PRAGMA_UNROLL_FULL)
-    #define _CCCL_PRAGMA_UNROLL_FULL() _Pragma("unroll")
-#endif // !defined(_CCCL_PRAGMA_UNROLL_FULL)
+#ifndef HIPCUB_PRAGMA_UNROLL_FULL
+    #define HIPCUB_PRAGMA_UNROLL_FULL() _Pragma("unroll")
+#endif // HIPCUB_PRAGMA_UNROLL_FULL
 
-#if !defined(_CCCL_PRAGMA_NOUNROLL)
-    #define _CCCL_PRAGMA_NOUNROLL() _Pragma("nounroll")
-#endif // !defined(_CCCL_PRAGMA_NOUNROLL)
+#ifndef HIPCUB_PRAGMA_NOUNROLL
+    #define HIPCUB_PRAGMA_NOUNROLL() _Pragma("nounroll")
+#endif // HIPCUB_PRAGMA_NOUNROLL
 
 #endif // HIPCUB_CONFIG_HPP_