+++ projects/hipcub/test/hipcub/test_hipcub_block_radix_rank.cpp
@@ -341,7 +341,7 @@ void rank_kernel(const KeyType* keys_input,
     UnsignedBits(&unsigned_keys)[ItemsPerThread]
         = reinterpret_cast<UnsignedBits(&)[ItemsPerThread]>(keys);
 
-    _CCCL_PRAGMA_UNROLL_FULL()
+    HIPCUB_PRAGMA_UNROLL_FULL()
     for(unsigned int key = 0; key < ItemsPerThread; key++)
     {
         unsigned_keys[key] = KeyTraits::TwiddleIn(unsigned_keys[key]);
@@ -592,7 +592,7 @@ void rank_with_prefix_sum_kernel(const KeyType* keys_input,
     UnsignedBits(&unsigned_keys)[ItemsPerThread]
         = reinterpret_cast<UnsignedBits(&)[ItemsPerThread]>(keys);
 
-    _CCCL_PRAGMA_UNROLL_FULL()
+    HIPCUB_PRAGMA_NOUNROLL()
     for(unsigned int key = 0; key < ItemsPerThread; key++)
     {
         unsigned_keys[key] = KeyTraits::TwiddleIn(unsigned_keys[key]);